upgrade trueusd to v3
