+++ .changeset/wicked-dots-work.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/trueusd-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
