--- packages/sources/trueusd-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as trueusd } from './trueusd'
