--- packages/sources/trueusd-test/test/integration/adapter.test.ts
@@ -1,65 +1,0 @@
-import {
-  TestAdapter,
-  setEnvVariables,
-} from '@chainlink/external-adapter-framework/util/testing-utils'
-import * as nock from 'nock'
-import { mockResponseFailure, mockResponseSuccess } from './fixtures'
-
-describe('execute', () => {
-  let spy: jest.SpyInstance
-  let testAdapter: TestAdapter
-  let oldEnv: NodeJS.ProcessEnv
-
-  beforeAll(async () => {
-    oldEnv = JSON.parse(JSON.stringify(process.env))
-    const mockDate = new Date('2001-01-01T11:11:11.111Z')
-    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-
-    const adapter = (await import('./../../src')).adapter
-    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
-      testAdapter: {} as TestAdapter<never>,
-    })
-  })
-
-  afterAll(async () => {
-    setEnvVariables(oldEnv)
-    await testAdapter.api.close()
-    nock.restore()
-    nock.cleanAll()
-    spy.mockRestore()
-  })
-
-  describe('trueusd endpoint', () => {
-    it('should return success', async () => {
-      mockResponseSuccess()
-      const response = await testAdapter.request({})
-      expect(response.statusCode).toBe(200)
-      expect(response.json()).toMatchSnapshot()
-      nock.cleanAll()
-    })
-
-    it('should return success when given a chain', async () => {
-      mockResponseSuccess()
-      const data = {
-        chain: 'AVA',
-        field: 'test',
-      }
-
-      const response = await testAdapter.request(data)
-      expect(response.statusCode).toBe(200)
-      expect(response.json()).toMatchSnapshot()
-      nock.cleanAll()
-    })
-
-    it('should return error when ripcord true', async () => {
-      mockResponseFailure()
-      const data = {
-        chain: 'AVA',
-      }
-      const response = await testAdapter.request(data)
-      expect(response.statusCode).toBe(502)
-      expect(response.json()).toMatchSnapshot()
-      nock.cleanAll()
-    })
-  })
-})
