--- packages/sources/trueusd/schemas/env.json
@@ -1,17 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/trueusd-adapter.json",
-  "title": "Chainlink External Adapter for TrueUSD",
-  "required": [],
-  "type": "object",
-  "properties": {
-    "API_ENDPOINT": {
-      "default": "https://core-api.real-time-attest.trustexplorer.io",
-      "type": "string"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
