+++ packages/sources/trueusd/src/config/index.ts
@@ -1,14 +1,9 @@
-import { Requester } from '@chainlink/ea-bootstrap'
-import { Config } from '@chainlink/ea-bootstrap'
+import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
 
-export const NAME = 'TRUEUSD'
-
-export const DEFAULT_ENDPOINT = 'trueusd'
-export const DEFAULT_BASE_URL = 'https://api.real-time-reserves.ledgerlens.io/v1/'
-
-export const makeConfig = (prefix?: string): Config => {
-  const config = Requester.getDefaultConfig(prefix)
-  config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
-  config.defaultEndpoint = DEFAULT_ENDPOINT
-  return config
-}
+export const config = new AdapterConfig({
+  API_ENDPOINT: {
+    description: 'API endpoint of adapter',
+    type: 'string',
+    default: 'https://api.real-time-reserves.ledgerlens.io/v1/',
+  },
+})
