+++ packages/sources/trueusd/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as TrueUsdInputParameters } from './trueusd'
-
-export type TInputParameters = TrueUsdInputParameters
-
-export * as trueusd from './trueusd'
+export { endpoint as trueusd } from './trueusd'
