+++ packages/sources/trueusd/src/index.ts
@@ -1,8 +1,13 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { endpointSelector, makeExecute } from './adapter'
-import { makeConfig, NAME } from './config'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { Adapter } from '@chainlink/external-adapter-framework/adapter'
+import { trueusd } from './endpoint'
+import { config } from './config'
 
-const adapterContext = { name: NAME }
+export const adapter = new Adapter({
+  defaultEndpoint: trueusd.name,
+  name: 'TRUEUSD',
+  config,
+  endpoints: [trueusd],
+})
 
-const { server } = expose(adapterContext, makeExecute(), undefined, endpointSelector)
-export { NAME, makeExecute, makeConfig, server }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
