--- packages/sources/trueusd/test/unit/trueusd.test.ts
@@ -1,27 +1,0 @@
-import { AdapterError, Requester } from '@chainlink/ea-bootstrap'
-import { assertError } from '@chainlink/ea-test-helpers'
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { makeExecute } from '../../src/adapter'
-
-describe('execute', () => {
-  const jobID = '1'
-  const execute = makeExecute()
-
-  describe('validation error', () => {
-    const requests = [
-      { name: 'invalid chain type', testData: { data: { chain: 1 } } },
-      { name: 'invalid field type', testData: { data: { field: 1 } } },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        try {
-          await execute(req.testData as unknown as AdapterRequest, {})
-        } catch (error) {
-          const errorResp = Requester.errored(jobID, error as AdapterError)
-          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
-        }
-      })
-    })
-  })
-})
